refactor: Decouple database and IPFS interactions from server setup
