+++ cmd/hway/main.go
@@ -10,9 +10,10 @@ import (
 	"github.com/labstack/echo/v4/middleware"
 	"github.com/onsonr/sonr/crypto/ucan"
 	"github.com/onsonr/sonr/internal/gateway"
-	"github.com/onsonr/sonr/pkg/common/ipfs"
 	config "github.com/onsonr/sonr/pkg/config/hway"
 	"github.com/onsonr/sonr/pkg/didauth/producer"
+	"github.com/onsonr/sonr/pkg/ipfsapi"
+	"gorm.io/gorm"
 )
 
 // main is the entry point for the application
@@ -23,6 +24,20 @@ func main() {
 		os.Exit(1)
 	}
 	os.Exit(0)
+}
+
+func initDeps(env config.Hway) (*gorm.DB, ipfsapi.Client, error) {
+	db, err := gateway.NewDB(env)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	ipc, err := ipfsapi.NewClient()
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return db, ipc, nil
 }
 
 func loadEnvImplFromArgs(args []string) (config.Hway, error) {
@@ -45,15 +60,7 @@ func loadEnvImplFromArgs(args []string) (config.Hway, error) {
 }
 
 // setupServer sets up the server
-func setupServer(env config.Hway) (*echo.Echo, error) {
-	ipc, err := ipfs.NewClient()
-	if err != nil {
-		return nil, err
-	}
-	db, err := gateway.NewDB(env)
-	if err != nil {
-		return nil, err
-	}
+func setupServer(env config.Hway, db *gorm.DB, ipc ipfsapi.Client) (*echo.Echo, error) {
 	e := echo.New()
 	e.Use(echoprometheus.NewMiddleware("hway"))
 	e.IPExtractor = echo.ExtractIPDirect()
