+++ pkg/didauth/producer/context.go
@@ -3,7 +3,7 @@ package producer
 import (
 	"github.com/labstack/echo/v4"
 	"github.com/onsonr/sonr/crypto/ucan"
-	"github.com/onsonr/sonr/pkg/common/ipfs"
+	"github.com/onsonr/sonr/pkg/ipfsapi"
 )
 
 type ProducerContext struct {
@@ -12,8 +12,8 @@ type ProducerContext struct {
 	TokenParser *ucan.TokenParser
 
 	// TokenStore is the token store used to store and retrieve tokens
-	TokenStore ipfs.IPFSTokenStore
+	TokenStore ipfsapi.IPFSTokenStore
 
 	// IPFSClient is the IPFS client used to resolve the UCAN
-	IPFSClient ipfs.Client
+	IPFSClient ipfsapi.Client
 }
