+++ pkg/didauth/producer/middleware.go
@@ -3,15 +3,15 @@ package producer
 import (
 	"github.com/onsonr/sonr/crypto/mpc"
 	"github.com/onsonr/sonr/crypto/ucan"
-	"github.com/onsonr/sonr/pkg/common/ipfs"
+	"github.com/onsonr/sonr/pkg/ipfsapi"
 
 	"github.com/labstack/echo/v4"
 )
 
 // Middleware returns middleware to spawn controllers and validate UCAN tokens
-func Middleware(ipc ipfs.Client, perms ucan.Permissions) echo.MiddlewareFunc {
+func Middleware(ipc ipfsapi.Client, perms ucan.Permissions) echo.MiddlewareFunc {
 	// Setup token store and parser
-	store := ipfs.NewIPFSTokenStore(ipc)
+	store := ipfsapi.NewIPFSTokenStore(ipc)
 	parser := ucan.NewTokenParser(perms.GetConstructor(), store, store)
 
 	// Return middleware
