+++ pkg/ipfsapi/client.go
@@ -1,4 +1,4 @@
-package ipfs
+package ipfsapi
 
 import (
 	"bytes"
